[Refactor] #9 dto 위치 변경
