+++ src/main/java/org/meetpl/recodingserver/api/search/controller/SearchController.java
@@ -2,14 +2,12 @@ package org.meetpl.recodingserver.api.search.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.meetpl.recodingserver.api.search.service.SearchService;
-import org.meetpl.recodingserver.api.search.service.dto.req.SearchReqDto;
-import org.meetpl.recodingserver.api.search.service.dto.res.SearchResDto;
+import org.meetpl.recodingserver.api.search.dto.req.SearchReqDto;
+import org.meetpl.recodingserver.api.search.dto.res.SearchResDto;
 import org.meetpl.recodingserver.global.common.SuccessResponse;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/search")
