+++ src/main/java/org/meetpl/recodingserver/api/search/service/SearchService.java
@@ -1,11 +1,12 @@
 package org.meetpl.recodingserver.api.search.service;
 
 import lombok.RequiredArgsConstructor;
-import org.meetpl.recodingserver.api.search.service.dto.req.SearchReqDto;
-import org.meetpl.recodingserver.api.search.service.dto.res.MemberSimpleResDto;
-import org.meetpl.recodingserver.api.search.service.dto.res.ReviewerInfoResDto;
-import org.meetpl.recodingserver.api.search.service.dto.res.SearchResDto;
+import org.meetpl.recodingserver.api.search.dto.req.SearchReqDto;
+import org.meetpl.recodingserver.api.search.dto.res.MemberSimpleResDto;
+import org.meetpl.recodingserver.api.search.dto.res.ReviewerInfoResDto;
+import org.meetpl.recodingserver.api.search.dto.res.SearchResDto;
 import org.meetpl.recodingserver.domain.member.service.MemberReader;
+import org.meetpl.recodingserver.domain.review.service.ReviewReader;
 import org.meetpl.recodingserver.domain.reviewer.domain.Reviewer;
 import org.meetpl.recodingserver.domain.reviewer.service.CareerReader;
 import org.meetpl.recodingserver.domain.reviewer.service.ReviewerReader;
@@ -15,11 +16,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class SearchService {
     private final ReviewerReader reviewerReader;
+    private final ReviewReader reviewReader;
     private final MemberReader memberReader;
     private final SkillReader skillReader;
     private final CareerReader careerReader;
@@ -38,8 +41,9 @@ public class SearchService {
                                 careerReader.getCareerByReviewerId(reviewer.getId())
                         ),
                         reviewer,
-                        skillReader.getSkillsByReviewerId(reviewer.getId())
+                        skillReader.getSkillsByReviewerId(reviewer.getId()),
+                        reviewReader.findReviewCountByReviewerId(reviewer.getId())
                 )
-        ).toList();
+        ).collect(Collectors.toList());
     }
 }
